moved roll related functions to their own folder
