+++ commands/rolls/index.js
@@ -1,0 +1,9 @@
+const { assignfreerolls } = require('./freerolls.js');
+const { giftrolls } = require('./giftrolls.js');
+const { roll } = require('./roll.js');
+
+module.exports = {
+    assignfreerolls,
+    giftrolls,
+    roll,
+}
